Correct handling of concrete array component types
The implementation of J9::ValuePropagation::isArrayCompTypeValueType
ended up assuming that if none of the special cases that it checked for
applied, the component type of the array might be a value type, and
returned TR_maybe as the result.  However, once all those special cases
are dealt with, the array must be a concrete type whose type is known
not to be a value type, and TR_no should be returned.

Also addressed some miscellaneous review comments.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>