Replace TR_RequireJITServer with a command line option
Passing `TR_RequireJITServer` environment variable on the client
will cause the JVM to exit with error, if it ever detects a
stream failure. This can be used for debugging purposes,
to ensure that a test fails if the server crashes, instead
of continuing compiling locally.
However, in real-life scenarios there are many false positives,
because if a client stream stops sending compilations for a time
and the stream times out (default timeout is 2 seconds),
we will detect a stream failure and terminate the JVM
even though the server is still running.
This commit replaces `TR_RequireJITServer` with a command line
option `-XX:+RequireJITServer` so that it can be more
easily passed to testing infrastructure and increases the default
timeout to 1 minute if the option is provided, while still respecting
user-specified timeouts.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>