Deprecate TR_DisableOOL option
This option is inconsistently used throughout the codebase. It was
originally introduced when OOL code sequences were first implemented as
a fallback path due to register allocation issues.

We have been using OOL seuqneces for almost a decade now and in some
places we don't even have a guard or fallback path against using OOL,
so enabling this option is not safe. Because of these reasons it is
best to just remove it.