Sending Client feedback about number of active compilations threads
When JITServer reaches 63 compilation threads any compilation request after that will be added
to the queue and requests waiting in the queue will experience delays and therefore the client can timeout.

In this commit, I send feedback to the client about the number of active compilation threads. I added a new
enum that represents the state of the active compilation threads "ServerActiveThreadsState", which has 3 states: HIGH_THREAD,
VERY_HIGH_THREAD and NORMAL_THREAD.

JITServer sends the client feedback about the state of Active threads at the end of every compilation. The client then
receives it and in the "JITServer::CompThreadActivationPolicy" decides based on the number of Active Threads when
to suspend/maintain the compilation threads.

If the number of Active threads >= to VERY_HIGH_THREAD, then we suspend active threads.
If the number of Active threads >= to HIGH_THREAD, then we maintain the number of active threads.
If the number of Active threads < HIGH_THREAD, then we subdue.

issue: #12459

Signed-off-by: Eman Elsabban <eman.elsaban1@gmail.com>