Avoid folding MutableCallSite.target
In the LambdaForm-based JSR292 implementation, CallSite declares a
target field to be inherited by all subclasses, including
MutableCallSite, and that field is declared final even though it's
actually mutable in instances of MCS.

Because it is a final field declared in a trusted JCL package, loads of
this field would normally be folded to a known object. However, folding
the target in the case of a MCS can result in the spurious removal of
the MCS target guard, so refuse to fold in that case.

Folding is prevented in transformIndirectLoadChain and
InterpreterEmulator.

This is a prerequisite to MutableCallSite inlining with LambdaForms.