Properly free ROMClass when caching at the JITServer fails
JITServer caches class information sent by the JVM clients in order to
reduce client-server network communication. This is done in a two step
operation: (1) First we allocate space for the ROMClass data structure
using persistent memory; (2) Then we acquire a lock and cache the class
information.
Due to a race condition, it is possible that, by the time we allocated
and initialized the ROMClass copy, another thread already inserted a copy
into the cache. If that happens, we just have to deallocate the newly
created ROMClass copy.
The bug solved by this commit revolves around how to the perform the
deallocation: if the server uses a global cache of ROMClasses (shared
among all the clients) then we have to use routines specific to this
global cache, rather than just calling persistentFree.

Fixes: #12701

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>