Improve JITServer::MessageBuffer capacity expansion mechanism
Message buffers store data that is exchanged between the server
and the client. When a message that is too big for the current buffer
is sent/received, `MessageBuffer::expandIfNeeded` or
`MessageBuffer::expand`
is called. These methods allocate a new, larger buffer and copy
any existing data over.

Previously, `MessageBuffer::expandIfNeeded` reallocated buffer to twice
the required size, and `MessageBuffer::expand` to exactly the required
size.
This leads to fragmentation of global persistent memory, as memory
blocks of arbitrary sizes cannot be easily reused for future allocations.

This commit sets the default buffer size to 32K, and all subsequent
expansions are done to the size of the nearest power of 2. This ensures
that memory allocations are standardized. Once a large buffer is freed,
subsequent large allocations will be able to reuse the
same memory block, while smaller allocations will split the block
into more reusable chunks, reducing fragmentation.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>