Refactor field shadow symbol creation
Deduplicate the logic that creates a field shadow (recognized or
unrecognized) with the final flag (and also the volatile and private
flags) set as appropriate.

Further, extract the logic to do so from a constant pool entry using
searchRecognizedField() and fieldAttributes() from InterpreterEmulator::
maintainStackForGetField(). This logic will be reused to create the
field symbol necessary to check whether folding should be avoided when
handling a KnownObjectTable_dereferenceKnownObjectField message.