Remove incorrect casting while calculating block frequency
In TR_BlockFrequencyInfo while calculating the rawCount of matched
blocks to the given bytecode info, it was incorrectly casting a 64-Bit
signed integer to uintptr_t. This conversion was forcing the function to
return very high frequency for the blocks for which we calculated
negative rawCount (cold blocks), which eventually causes compiler to
optimize cold code leaving with sub-optimal compiled code.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>