Remove TR_DebugExt
TR_DebugExt is a facility that allowed one to debug the compiler
(through a coredump file) using the compiler. However, it is not
actively maintained, and is not very stable (fundamentally because
rather than using debug symbols, it tries to executing TR code to print
the C/C++ structures). This commit removes this facility.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>