Fix -Xjit:classesWithFoldableFinalFields={...}
foldFinalFieldsIn() was previously passing className directly to
TR::SimpleRegex::match(), but className is not necessarily
NUL-terminated. As a result, classes named in the pattern could fail to
match due to trailing garbage following the name. It was possible to
work around this problem by specifying patterns that end with a
wildcard, e.g. {Foo*} instead of {Foo}.