Process a list of test exclude files
Currently, IncludeExcludeTestAnnotationTransformer only processes a single
exclude file.

It has been updated to process a list of exclude files.

The list can be provided via the "EXCLUDE_FILE" environment variable where the
exclude files are separated by commas.

The name of the environment variable is left unchanged. So, the existing
scripts do not need to be updated.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>