Fix incorrect J9Class pointer in ClassCastException
When running 64 bit compressed refs, if clazz is an address
above the 2G boundary then we can't use a push 32bit immediate
instruction to pass it on the stack to the jitThrowClassCastException
helper as the address gets sign extended. It needs to be
stored in a temp register and then push the register to the stack.

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>