Fix a bug in ResolvedMethod_getMultipleResolvedMethods message handler
The client-side handler of `ResolvedMethod_getMultipleResolvedMethods`
did not check the value of `J9JIT_RUNTIME_RESOLVE` flag for some of
the methods it created, which resulted in bugs when `-Xjit:rtResolve`
option is specified.

This commit refactors `ResolvedMethod_getMultipleResolvedMethods`
and `TR_ResolvedJ9JITServerMethod::cacheResolvedMethodCallees` to
call resolved method creation routine on the client directly, instead
of copy-pasting the code from that routine, which fixes the problem.

The code here is a small chunk of PR #12147, which should fully refactor
resolved methods creation, when it is merged at a later date.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>