Update ClassLoader_defineClassImpl1
In OJDK MethodHandles.Lookup, defineClassImpl1 is used to create both
hidden and normal classes.

Only for hidden classes the STRONG and NESTMATE flags are used.

Also, the current code will set the ANON flag for a normal class if the STRONG
flag is missing.

The code has been reorganized so that normal classes are created without the
ANON flag.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>