Fix Java options for testDDRExt_SharedClasses
For the test testDDRExt_SharedClasses, add -Xmx512M to the TCK.cache.initialization
step to match the TCK.generate.dump step. This ensures that the SCC AOT header
matches the SCC AOT contents, and that the AOT code in the SCC is not ignored due
to a compressedrefs shift mismatch.

Signed-off-by: midronij <jackie.midroni@ibm.com>