Fix client-side mirror of method handle archetype specimen
`TR_J9ServerVM::createMethodHandleArchetypeSpecimen` creates a resolved
method, flags it as archetype specimen and sets method handle location.
The problem with previous implementation is that it first creates a
client-side mirror and then sets the above flag on the server-side.

However, at a later point, client might check the value of the archetype
specimen flag, e.g. when `TR_ResolvedJ9Method::getResolvedStaticMethod`
is called. This method uses it to determine whether a static method must
be resolved at compile time, even if the runtime resolved flag is specified.
Since the value won't be set, the method will remain unresolved, leading
to crashes.

This commit changes the code to set the archetype specimen flag and
method handle location on both server and client at resolved method
creation.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>