Increase maximum number of JITServer compilation threads
The new limit is 999 (1000 including the statistics/diagnostic
thread) and is somewhat artificial. It allows us to keep the
existing format of compilation thread names defined in
TR::CompilationInfoPerThread::CompilationInfoPerThread(), and is
large enough for a single machine, at least in the near future.

The default number of threads is kept at 63. If the user
specifies (with -XcompilationThreads<N>) a number of threads
larger than the limit of 999, a diagnostic message is printed
to stderr, and the default number of threads (63) is used.

If the requested number of threads is larger than the maximum
number of code caches, a diagnostic message is printed to
stderr, and JITServer start is aborted.

Signed-off-by: Alexey Khrabrov <khrabrov@cs.toronto.edu>