Fix bug in assumption reclamation
`TR_RuntimeAssumptionTable::reclaimAssumptions` marks assumptions
for detach and sets the head of the assumption list to NULL.
However, it has an issue where it does not guarantee that
`metaData->runtimeAssumptionList` will be set to NULL,
since the sentinel double pointer passed to it can be the address
of the list in the compilation object, instead of in the metadata.
It's also not always the case that the method is passed metadata
pointer when it exists.

This commit sets `_owningMetadata` attribute of a sentinel runtime
assumption at the points metadata is created.
At reclamation, we use the owning metadata to reset the assumption
list pointer, if metadata pointer is not directly available

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>