Fix a memory leak in runtime assumption list
The mark-and-sweep assumption reclamation algorithm will free
marked for detach entries in
`TR_RuntimeAssumptionTable::reclaimMarkedAssumptionsFromRAT`.
The method iterates through the runtime assumption table and
frees all entries that are marked for detach, after dequeing
them from the per-compilation runtime assumption list.

However, sentinel runtime assumptions (i.e. heads of assumption lists)
are detached but never freed, since they are not part of the RAT.
The fix to this memory leak is to delete sentinels the moment they
are dequeued from the runtime assumption list.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>