Skip ArrayStoreCHK or BNDCHK in VP if method allows for that
Allow Value Propagation to skip ArrayStoreCHKs or BNDCHKs for calls to
jitStoreFlattenableArrayElement that it transforms if the owning method
of the call node indicates they can be skipped.  Similarly, allow
BNDCHKs to be skipped for calls to jitLoadFlattenableArrayElement that
are transformed.

This mirrors knowledge about the JCL that IL generation takes advantage
of as it generates IL for array operations.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>