Fix and re-enable inliner ILGen reduction of getCallerClass
The condition and value `J9_EXTENDED_RUNTIME_ALLOW_GET_CALLER_CLASS`
at the start of the current implementation causes us to disable this
optimization by default. Digging into the history it seems this was an
inadvertent change when the `getStackClass(int)` overload was removed
and deprecated from the class library.

We reimplement and fix existing bugs in the implementation of this
reduction and enable it by default. We also provide an escape hatch in
the form of an environment variable should we encoutner corner case bugs
within this implementation.

The implementation of this reduction follows closely the implementation
in the VM found in `getCallerClassJEP176Iterator` which we used as
reference. The JIT code in addition has to handle not being able to see
past the toplevel method being compiled.