z/OS jenkins build: Add a target to build JTreg test libraries
Some of openjdk's JTreg tests falling under the `jdk` category also include
native libraries. These libraries need to be built for z/OS too, and be made
available in the test-images.tar.gz artifact generated by the builds. The
makefiles that build, link and copy these libraries have been suitably modified
on the z/OS repository. This pull request proposes the addition of the
`test-image-jdk-jtreg-native` target for z/OS JDK11 builds, to trigger
these makefile changes.

Please note that we cannot add the `test-images` target because it also
includes `test-image-hotspot-jtreg-native` - a target for building Hotspot
natives - which we wouldn't want to include for openj9.

Signed-off-by: Pushkar N Kulkarni <pushkar.nk@in.ibm.com>