Add an option to log JITServer connections
This commit adds a new command line option `-XX:+LogJITServerConns`
which enables the newly added verbose option `TR_VerboseJITServerConns`.
The option can be enabled on both server and client.
When enabled, it logs connections/disconnections between server and client.

This commit also adds UID for the server. The client stores the latest
server UID to determine when a connection to a new server occurs.

Closes: #13055

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>