Introduce portable and non-portable CRIU modes
When CRIU support is enabled via `-XX:+EnableCRIUSupport` there are two
phases the JVM can be in. Portable or Non-portable phase.
Portable vs Non-portable phase is determined by the result of calling
`isCheckpointAllowed(..)`. If the result is TRU then it is in portable
phase, non-portable otherwise.

THe JVM can only be checkpointed if it is in the portable phase. If
`-XX:+CRIURestoreNonPortableMode` is specified as
soon as the JVM is restored it transitions to non-portable phase.

This behaviour notifies the JIT on when to be more conservative in
regards to optimizations that depend on environment properties and when
the JIT is free to assume that the environment is stable.

The default behaviour is portable phase all the time
(-XX:-CRIURestoreNonPortableMode). This behaviour comes with a
performance penalty.

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>