Fix incompatible arguments for inlining method from hidden class
While inlining non-static method calls, we build the parameters where
the very first slot of the parameters is the class of which that method
belongs to. We later build ArgInfo from the method symbol of the callee
by iterating through this list of parameters and for any parameter which
is Class, we query VM to get the class pointer from the signature of
the class. This was causing an issue when we were trying to inline a
method that belongs to hidden class as VM would return NULL when queried
to get class from signature.

We do not need to call VM to get the class pointer for the
first parameter slot of the non-static method calls as we have that
information avaible through resolvedMethodSymbol.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>