Use jitSetMutableCallSiteTarget for all instances of MutableCallSite
...including in particular instances of subclasses. Previously,
setCallSiteTargetImpl() would recognize an instance as an MCS only if
its concrete type was exactly MutableCallSite. As a result, if the
target of an instance of a subclass of MutableCallSite were inlined, and
if that call site were later updated with a different target, we would
fail to invalidate the runtime assumptions.