Replace MethodHandleImpl.profileBoolean() calls with boolean parameter
java/lang/invoke/MethodHandleImpl.profileBoolean() performs some internal profiling
on the boolean parameter value before returning it.  Since the OpenJ9 JIT does not
consume that profiling information the profiling overhead is not necessary and
interferes with inlining decisions..  Eliminate the call and simply replace it with
a reference to the boolean parameter.

Add java/lang/invoke/MethodHandleImpl.isCompileConstant() as a future recognized
method.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>