Fix off-by-one error in String.indexOf acceleration on z15
- `stringIndexReg` contains the current index into the search string
  that we will start the search on
- `maxIndexReg = searchString.length() - patternString.length()`
  and this value does not change at all
- `loadLenReg` at this point in code is always 16

The problem here is that due to the comparison we are making we break
the assumption that the remaining number of characters we have left to
examine is always greater than or equal to the length of the pattern.
The code snippet being changed has two branches:

- The first branch branches if `stringIndexReg >= maxIndexReg`
- The second branch branches if `stringIndexReg >= searchString.length()`

There are two problems with the first branch. First problem is that
`stringIndexReg` is updated after the branch, when it should be updated
before the branch. Second issue is that we can have a false negative
result because we branch on equality when we really shouldn't.