Transform Unsafe.allocateInstance() calls into new during VP
...in cases where the java/lang/Class argument is a known object. The
java/lang/Class must be known to be non-null, and the class must be
already initialized. This transformation avoids a J2I transition.

This change adds the class initialization test to the transformation of
Object.newInstancePrototype() as well. Object.newInstancePrototype() is
part of the implementation of Class.newInstance(), which is supposed to
trigger initialization if necessary.