Patch HiddenClass's moduleObject initialization
A hiddenClass should have the same defining classLoader/runtime package
as the Loopkup class (ie. hostClass). so VM need to use the hostClassLoader
when setting the unnamed module for hiddenClass as VM may uses the
anonymousClassLoader to load hiddenClass and update to hostClassLoader
after class load. This will cause hiddenClass to have different module
than the hostClass.

Signed-off-by: Jack Lu <Jack.S.Lu@ibm.com>