Define an objectInequalityComparison non-helper to compare references
The jitAcmpneHelper helper has been introduced to make reference
inequality comparison a first-class operation when value type support
is enabled, just as reference equality comparison is.  This change
introduces a corresponding <objectInequalityComparison> non-helper to
represent that operation which ultimately relies on the jitAcmpneHelper.

This also changes IL Generation to use the <objectInequalityComparison>
non-helper instead of <objectEqualityComparison>.  That allows IL
Generation to use an ificmpeq of the result with zero in translating the
if_acmpeq bytecode, and an ificmpne of the result with zero in
translating the if_acmpne bytecode.

In contrast, if IL Generation uses <objectEqualityComparison>, it
compares the result using ificmpne for an if_acmpeq, and ificmpeq for an
if_acmpne.  Preserving the sense of the comparison ensures that
TR_IProfiler::getBranchCounters assigns profiling frequencies correctly.

Fixes issue #13663

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>