Do not reuse fabricated symrefs for relocatable compilations
The OpenJDK method handle code uses `findOrFabricateShadowSymbol`
to create symbol references needed for recognized call transformer.
The routine allows reuse of symbol references for recognized fields
even for non-private fields that belong to different classes.
This works fine in a regular compilation, however in a remote or AOT
compilation, the constant pool of fields reused outside its class
will not be relocated correctly, leading to crashes.

This commit adds a check to disallow reuse of fabricated shadow symbols
in JITServer/AOT compilations.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>