Fix the race condition timing hole for classPathEntries and cpEntryCount
The current code records the classPathEntries and pass it along with the
latest cpEntryCount to findLocallyDefinedClassFunction. The
classPathEntries and cpEntryCount could be changed in between by
addJarToSystemClassLoaderClassPathEntries(), creating
a timing hole for race condition. Also it is possible that a thread
could read the stale classPathEntries which is freed by
addJarToSystemClassLoaderClassPathEntries().

Change classloader->classPathEntries to a pointer array. 

For bootstrap class loader, its class path entries can only be appended,
not removed. So the J9ClassPathEntry is never freed. The pointer array
is incremented by size of 64 when necessary. 

issue #13414

Signed-off-by: Hang Shao <hangshao@ca.ibm.com>