Refine linkToVirtual to a direct call when no vtable entry exists
MethodHandleTransformer refines linkToVirtual calls to the target
method when MemberName is known. The call kind of the refined
linkToVirtual call target was always treated as Virtual.
However, for private virtual methods and java/lang/Object;
type virtual methods, there is no corresponding entry in the vtable,
and for such methods the interpreter vtable index is set as 0.
The dispatch is not performed through the vtable entry, but instead
it is a direct call to the target method obtained from MemberName.

By checking if jitVtableOffset is positive, we can determine if
a call needs to be dispatched through the vTable. Otherwise, we
can handle it similar to linkToStatic.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>