Allow downgrading of JSR292 methods
To improve start-up/ramp-up of applications, OpenJ9 allows some compilations
to be downgraded from "warm" (the default optimization level) to "cold"
optimization level. However, current logic prevents downgrading of methods
that are flagged as MethodHandleInvokes because inlining is crucial for
the performance of such methods.
In order to improve start-up time of applications, this commit allows the
downgrading of JSR292 methods under certain conditions:
1. The JVM is in the start-up phase
2. Shared classes are enabled explicitly on the command line
3. AOT is enabled

The reason is that, if the shared classes are explicitly enabled on the
command line and AOT is not disabled, there is a good likelihood that the
user cares about start-up time.
Performance evaluation has shown that start-up time of Liberty+DT7 can be
improved by 5-6% with Java17 and by ~1% with Java11. There was no throughput
impact from these changes.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>