Return uintptr_t from vTableOrITableIndexFromMemberName()
The result is the value of J9JNIMethodID::vTableIndex, which is UDATA.
The high bit of this value is a flag that distinguishes VFT offsets from
iTable indices. The previous return type (int32_t) required this value
to be truncated on 64-bit platforms, losing the value of the flag,
leaving callers with no way to distinguish the cases.