Avoid refining linkToVirtual to interface method in InterpreterEmulator
vTableOrITableIndexFromMemberName() returns either a VFT offset or an
iTable index tagged with J9_JNI_MID_INTERFACE. In the latter case,
treating it as a VFT offset and generating a virtual call is incorrect.
We could determine the interface callee and replace linkToVirtual() with
an interface call, but interface call sites currently require a CP
index in order to find call targets.

Additionally, vTableOrITableIndexFromMemberName() can return zero to
indicate that the method should be dispached directly. This case is now
handled properly by setting isIndirectCall to false, which results in
the creation of a direct call site. Previously we would get a virtual
call site, which would find the correct (nonoverridden) target method,
but that method would not be inlined with the correct guard (in
particular, an HCR guard in HCR mode, which is the default).