Avoid refining linkToVirtual to interface method in MH transformer
vTableOrITableIndexFromMemberName() returns either a VFT offset or an
iTable index tagged with J9_JNI_MID_INTERFACE. In the latter case,
treating it as a VFT offset and generating a virtual call is incorrect.
In case an attempt to do so results in a positive JIT VFT offset and
consequently a direct call, that is also incorrect. We could determine
the interface callee and replace linkToVirtual() with an interface call,
but interface call nodes must currently be unresolved. Code generation
requires a CP index in the dispatch.