Workaround for zlinux mmap issue in pltest 0.28.0
On s390x Ubuntu 20.04, if a write attempt is made on mmap'd memory whose
protection is set to read only, and then a second attempt is made when
the map protection is set to back to writeable, the second attempt will
not update the underlying memory.
This workaround does a small read so that the write attempt on writeable
memory will work.

Signed-off-by: Mike Zhang <mike.h.zhang@ibm.com>