Fix memcpy byte-to-char loop reduction for little endian architectures
Idiom recognition is incorrectly transforming byte-to-char loops for little
endian.  For example,
```
   for (int i = 0, j = 0; j < numChars; i += 2, j++) {
      dst[j + offsetDst] = (char) (((src[i + 1 + offsetSrc] & 0xFF) << 8) | (src[i + offsetSrc] & 0xFF));
   }
```
A loop such as this can be recognized and reduced to an arraycopy on little endian
architectures.  Similarly, if the char was composed with the shift on the lower
address byte then it would be recognized as an arraycopy on big endian architectures.

However, for little endian, the wrong address tree is chosen to start the arraycopy
(i.e., `src[i + 1 + offsetSrc]` is chosen instead of  `src[i + offsetSrc]`).  Big
endian selects the correct address.

The logic that determines this is conditional on the array element store being
unsigned, which would have been true prior to the retirement of `cstore` IL opcodes.
Now that character stores are implemented with `sstore`s the unsigned check should
be removed.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>