Notify JIT of class redefinitions that used HCR extensions
Previously, if a class was redefined using extensions (i.e. a method
got added/removed to/from the class), JIT event data was not populated
and the JIT class redefinition hook would receive an empty class list.
This works for regular JIT, since the entire code cache is discarded
after such redefinition, this information is not needed on the JIT side.

However, this does not work correctly for JITServer, because if a
client JIT receives an empty class list, the server will not get notified
that a class got redefined, and might use invalid cached information in
the future compilations.

This commit changes the VM code to pass a variable to the class
redefinition hook that indicates whether extensions were used.
The hook will then add ~0 to the list of JITServer client unloaded
classes.
Once a server receives it, it will clear all the caches.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>