Remove InterfaceHandle.interfaceCall_*
It's not possible yet to refine calls to these methods, and as such the
thunk archetypes for InterfaceHandle don't use them. Just remove them
since this InterfaceHandle is on the way out along with the rest of the
OpenJ9 MethodHandle implementation.

This removal simplifies the IL generator and InterpreterEmulator logic
slightly in preparation for a few bug fixes.