Extract logic that determines the method of a Direct- or VirtualHandle
This logic was duplicated in two locations, and in each of those
locations duplicated between in-process and JITServer compilation paths,
for a total of 4 copies.

By calling into the new method, refineResolvedCalleeForInvokestatic()
will now hold VM access while the java/lang/Class object is
dereferenced. Previously, it was possible for the object to move while
the compiler was concurrently trying to dereference it.

refineResolvedCalleeForInvokestatic() no longer inserts the MethodHandle
into the known object table. The index was used only for tracing. If for
some reason the object is not already present, then it will be inserted
later at the first use of getOrCreateIndex() whose result is needed.