Fix crash when VirtualHandle.virtualCall_* calls an interface method
It's possible for the method represented by a VirtualHandle to be an
interface method when the (non-interface) class passed to findVirtual()
inherits a method from an interface. The inherited method could be a
default method, or it could be an abstract interface method if the class
is also abstract.

Prior to this change, when the IL generator or the inliner was
processing a call to virtualCall_* and looking to rewrite it in terms of
the represented method, the compiler would look for the J9Method in the
vTable of the class specified by the handle's defc field. This is the
defining class of the represented method, so in case of an interface
method, defc is an interface. An interface has no vTable, so the
J9Method pointer would be the result of an out-of-bounds load. This bad
value would later cause a crash in the compiler.

Now the method is looked up instead in the vTable of the class
identified by the handle's referenceClass field, which is the class that
was originally provided to findVirtual(). The vTable offset in vmSlot is
correct for this class.

Finally, InterpreterEmulator will now refuse to refine virtualCall_*
when calling an interface method in inliner. There's no way to get the
vTable offset or the corresponding class based on the interface method
alone. While it should be possible to plumb the required information out
into visitInvokestatic(), doing so would only benefit OpenJ9 method
handles, which are on the way out. The analogous linkToVirtual() (for
OpenJDK method handles) calls the method specified by a MemberName
object, which does not also specify the class originally passed to
findVirtual(), so it will not be possible to determine the vTable
offset.